requirements.txt
asgiref==3.4.1
Django==3.2.18
django-phonenumber-field==6.0.0
phonenumbers==8.13.11
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.1.1
